refactor: Move headerstoRecords function to separate file
The code changes in `ServerController.ts` move the `headerstoRecords` function to a separate file `headerstoRecords.ts` for better organization and reusability. This improves the maintainability of the codebase and follows best practices for modularization.
