+++ apps/server/project.json
@@ -1,0 +1,3 @@
+{
+  "projectType": "application"
+}
