+++ apps/server/test/headerstoRecords.test.ts
@@ -1,0 +1,35 @@
+import { describe, expect, test } from "vitest";
+import { headerstoRecords } from "../src/headerstoRecords.js";
+import type { IncomingHttpHeaders } from "http";
+
+describe("headerstoRecords", () => {
+  test("should convert headers to records", () => {
+    const headers: IncomingHttpHeaders = {
+      "content-type": "application/json",
+      "x-auth-token": "abc123",
+      accept: "application/json, text/html",
+    };
+
+    const expected = {
+      "content-type": "application/json",
+      "x-auth-token": "abc123",
+      accept: "application/json, text/html",
+    };
+
+    const result = headerstoRecords(headers);
+
+    expect(result).toEqual(expected);
+  });
+
+  test("should handle empty headers", () => {
+    const headers = {};
+
+    const expected = {};
+
+    const result = headerstoRecords(headers);
+
+    expect(result).toEqual(expected);
+  });
+
+  // Add more test cases as needed
+});
