--- apps/server/test/index.test.ts
@@ -1,7 +1,0 @@
-import { describe, test, expect } from "vitest";
-
-describe("example", () => {
-  test("example", () => {
-    expect(1).toBe(1);
-  });
-});
